Removes a couple requirements from the cpanfile, as well as templates that correspond to those requirements.
Edits Dockerfile.cpan such that only the cpanfile is copied over.

Edits the cpanfile such that all 'this version or higher' requirements
are changed to 'this version only', which satisfies the reasoning for
not copying the cpanfile.snapshot over in Dockerfile.cpan (as far as I
understand).

Creates a shell script for building Dockerfile.cpan and copying over the
resulting cpanfile.snapshot.